+++ packages/api/src/templates/index.ts
@@ -1,0 +1,6 @@
+export { default as successEmail } from "./success-email";
+
+export interface Template {
+  html: string;
+  text: string;
+}