+++ packages/app/src/components/MinistereTravail/SearchButton.tsx
@@ -1,0 +1,17 @@
+/** @jsx jsx */
+import {FC, MouseEventHandler} from "react";
+import {IconSearch} from "../Icons";
+import {jsx} from "@emotion/core";
+import Button from "./Button";
+
+interface SearchButtonProps {
+  onClick: MouseEventHandler
+}
+
+const SearchButton: FC<SearchButtonProps> = ({ onClick }) => (
+  <Button onClick={onClick}>
+    <IconSearch/>
+  </Button>
+);
+
+export default SearchButton;