+++ packages/app/src/components/MinistereTravail/colors.tsx
@@ -1,0 +1,5 @@
+export const PINK = "#EA148C";
+export const DARK_PINK = "#B51F73";
+export const TITLE = "#3C4046";
+export const TEXT = "#000000";
+export const GREEN = "#91ae4f";